pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "jarvis-v2"
version = "0.1.0"
description = "FiestyGoat AI — LangGraph-native autonomous revenue system"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=1.0.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-ollama>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "redis>=5.2.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.0",
    "neo4j>=5.26.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "structlog>=24.4.0",
    "alpaca-trade-api>=3.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "black>=24.10.0",
    "ruff>=0.7.0",
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
